增删改查 规格组判断删除

+++ mingrui-shop-service/mingrui-shop-service-xxx/src/main/java/com/baidu/shop/service/impl/SpecGroupServiceImpl.java
@@ -3,9 +3,12 @@ package com.baidu.shop.service.impl;
 import com.alibaba.fastjson.JSONObject;
 import com.baidu.shop.base.BaseApiService;
 import com.baidu.shop.base.Result;
+import com.baidu.shop.dto.SpecParamDTO;
 import com.baidu.shop.dto.SpecificationDTO;
+import com.baidu.shop.entity.SpecParamEntity;
 import com.baidu.shop.entity.SpecificationEntity;
 import com.baidu.shop.mapper.SpecGroupMapper;
+import com.baidu.shop.mapper.SpecParamMapper;
 import com.baidu.shop.service.SpecificationService;
 import com.baidu.shop.utils.BaiduBeanUtil;
 import org.springframework.web.bind.annotation.RestController;
@@ -26,6 +29,9 @@ public class SpecGroupServiceImpl extends BaseApiService implements Specificatio
 
     @Resource
     private SpecGroupMapper specGroupMapper;
+
+    @Resource
+    private SpecParamMapper specParamMapper;
 
     @Override
     public Result<List<SpecificationEntity>> listSpecGroup(SpecificationDTO specificationDTO) {
@@ -50,7 +56,42 @@ public class SpecGroupServiceImpl extends BaseApiService implements Specificatio
 
     @Override
     public Result<JSONObject> delSpecGroup(Integer id) {
+        //删除规格组之前需要先判断一下当前规格组下是否有规格参数
+        //true : 不能被删除
+        //false -->
+        Example example = new Example(SpecParamEntity.class);
+        example.createCriteria().andEqualTo("groupId",id);
+        List<SpecParamEntity> specParamEntities = specParamMapper.selectByExample(example);
+        if(specParamEntities.size() != 0) return this.setResultError("不能删除");
+
         specGroupMapper.deleteByPrimaryKey(id);
+        return this.setResultSuccess();
+    }
+
+    @Override
+    public Result<List<SpecParamEntity>> getSpecParam(SpecParamDTO specParamDTO) {
+        SpecParamEntity specParamEntity = BaiduBeanUtil.copyProperties(specParamDTO, SpecParamEntity.class);
+        Example example = new Example(SpecParamEntity.class);
+        example.createCriteria().andEqualTo("groupId",specParamEntity.getGroupId());
+        List<SpecParamEntity> specParamEntities = specParamMapper.selectByExample(example);
+        return this.setResultSuccess(specParamEntities);
+    }
+
+    @Override
+    public Result<JSONObject> saveSpecParam(SpecParamDTO specParamDTO) {
+        specParamMapper.insertSelective(BaiduBeanUtil.copyProperties(specParamDTO,SpecParamEntity.class));
+        return this.setResultSuccess();
+    }
+
+    @Override
+    public Result<JSONObject> editSpecParam(SpecParamDTO specParamDTO) {
+        specParamMapper.updateByPrimaryKeySelective(BaiduBeanUtil.copyProperties(specParamDTO,SpecParamEntity.class));
+        return this.setResultSuccess();
+    }
+
+    @Override
+    public Result<JSONObject> delSpecParam(Integer id) {
+        specParamMapper.deleteByPrimaryKey(id);
         return this.setResultSuccess();
     }
